project managemet tool/requirements.txt
Django>=4.2
djangorestframework
python-dateutil
Pillow